+++ compiler-core/src/javascript/tests.rs
@@ -75,8 +75,8 @@ macro_rules! assert_js {
         // TODO: Currently we do this here and also in the tests. It would be better
         // to have one place where we create all this required state for use in each
         // place.
-        let _ = modules.insert("gleam".to_string(), crate::type_::build_prelude(&ids));
-        let (mut ast, _) = crate::parse::parse_module($dep_src).expect("dep syntax error");
+        let _ = modules.insert("gleam".to_string(), $crate::type_::build_prelude(&ids));
+        let (mut ast, _) = $crate::parse::parse_module($dep_src).expect("dep syntax error");
         ast.name = $dep_name;
         let dep = crate::type_::infer_module(
             crate::build::Target::JavaScript,
@@ -216,8 +216,8 @@ macro_rules! assert_ts_def {
         // TODO: Currently we do this here and also in the tests. It would be better
         // to have one place where we create all this required state for use in each
         // place.
-        let _ = modules.insert("gleam".to_string(), crate::type_::build_prelude(&ids));
-        let (mut ast, _) = crate::parse::parse_module($dep_src).expect("dep syntax error");
+        let _ = modules.insert("gleam".to_string(), $crate::type_::build_prelude(&ids));
+        let (mut ast, _) = $crate::parse::parse_module($dep_src).expect("dep syntax error");
         ast.name = $dep_name;
         let dep = crate::type_::infer_module(
             crate::build::Target::JavaScript,