+++ compiler-core/src/type_/tests.rs
@@ -297,7 +297,7 @@ macro_rules! assert_with_module_error {
         // place.
         let _ = modules.insert("gleam".to_string(), build_prelude(&ids));
         // Repeatedly create importable modules for each one given
-        let (mut ast, _) = crate::parse::parse_module($module_src).expect("syntax error");
+        let (mut ast, _) = $crate::parse::parse_module($module_src).expect("syntax error");
         ast.name = $name;
         let module = infer_module(
             Target::Erlang,
@@ -311,7 +311,7 @@ macro_rules! assert_with_module_error {
         .expect("should successfully infer");
         let _ = modules.insert($name.join("/"), module.type_info);
 
-        let (mut ast2, _) = crate::parse::parse_module($module_src2).expect("syntax error");
+        let (mut ast2, _) = $crate::parse::parse_module($module_src2).expect("syntax error");
         ast2.name = $name2;
         let module = infer_module(
             Target::Erlang,
@@ -383,7 +383,7 @@ macro_rules! assert_warning {
         insta::assert_snapshot!(insta::internals::AutoName, output, $src);
     };
     ($src:expr, $warning:expr $(,)?) => {
-        let (mut ast, _) = crate::parse::parse_module($src).expect("syntax error");
+        let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = vec!["my_module".to_string()];
         let mut warnings = vec![];
         let ids = UniqueIdGenerator::new();
@@ -418,7 +418,7 @@ macro_rules! assert_warning {
         let _ = modules.insert("gleam".to_string(), build_prelude(&ids));
         // Repeatedly create importable modules for each one given
         $(
-        let (mut ast, _) = crate::parse::parse_module($module_src).expect("syntax error");
+        let (mut ast, _) = $crate::parse::parse_module($module_src).expect("syntax error");
         ast.name = $name;
         let module = infer_module(
             Target::Erlang,
@@ -490,7 +490,7 @@ macro_rules! assert_no_warnings {
         let _ = modules.insert("gleam".to_string(), build_prelude(&ids));
         // Repeatedly create importable modules for each one given
         $(
-        let (mut ast, _) = crate::parse::parse_module($module_src).expect("syntax error");
+        let (mut ast, _) = $crate::parse::parse_module($module_src).expect("syntax error");
         ast.name = $name;
         let module = infer_module(
             Target::Erlang,
@@ -505,7 +505,7 @@ macro_rules! assert_no_warnings {
         let _ = modules.insert($name.join("/"), module.type_info);
         )*
 
-        let (mut ast, _) = crate::parse::parse_module($src).expect("syntax error");
+        let (mut ast, _) = $crate::parse::parse_module($src).expect("syntax error");
         ast.name = vec!["my_module".to_string()];
         let _ = infer_module(
             Target::Erlang,
@@ -564,16 +564,19 @@ fn field_map_reorder_test() {
         fields: HashMap::new(),
         args: vec![
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: None,
                 value: int("1"),
             },
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: None,
                 value: int("2"),
             },
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: None,
                 value: int("3"),
@@ -582,16 +585,19 @@ fn field_map_reorder_test() {
         expected_result: Ok(()),
         expected_args: vec![
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: None,
                 value: int("1"),
             },
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: None,
                 value: int("2"),
             },
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: None,
                 value: int("3"),
@@ -605,16 +611,19 @@ fn field_map_reorder_test() {
         fields: [("last".to_string(), 2)].into(),
         args: vec![
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: None,
                 value: int("1"),
             },
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: None,
                 value: int("2"),
             },
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: Some("last".to_string()),
                 value: int("3"),
@@ -623,16 +632,19 @@ fn field_map_reorder_test() {
         expected_result: Ok(()),
         expected_args: vec![
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: None,
                 value: int("1"),
             },
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: None,
                 value: int("2"),
             },
             CallArg {
+                implicit: false,
                 location: Default::default(),
                 label: Some("last".to_string()),
                 value: int("3"),