+++ CHANGELOG.md
@@ -1,14 +1,23 @@
 # Changelog
 
-# Unreleased
+## Unreleased
 
 - Generated HTML documentation now includes the `theme-color` HTML meta tag.
 - The `use` expression has been introduced. This is a new syntactic sugar that
   permits callback using code to be written without indentation.
+- Nightly builds are now also published as OCI container images hosted on
+  GitHub.
+- Fixed a bug where the build tool would not hook up stdin for Gleam programs it
+  starts.
+- Fixed a bug where using a record constructor as a value could generate a
+  warning in Erlang.
 - Fixed a bug where the warning for `todo` would not print the type of the code
   to complete.
 - Changed float size limit in bitstring expressions to 16, 32 or 64, when static.
   Also allowed dynamic size.
+- Generated HTML documentation now includes all static assets (but the web
+  fonts), so that it can be accessed offline or in far future once CDNs would
+  404.
 
 ## 0.24.0 - 2022-10-25